# swagger.yaml
openapi: 3.0.3
info:
  title: MEDAPP API
  description: MEDAPP API
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local server
...
paths:
...
  /api/v1/nurses:
    get:
      tags:
        - Nurses
      summary: Get all nurses
      operationId: getAllNurses
      responses:
        "200":
          description: List of nurses
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Nurse"
    post:
      tags:
        - Nurses
      summary: Create a new nurse
      operationId: createNurse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNurseRequest"
      responses:
        "201":
          description: Nurse created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Nurse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
  /api/v1/nurses/{id}:
    get:
      tags:
        - Nurses
      summary: Get nurse by ID
      operationId: getNurseByUserId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nurse details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Nurse"
        "404":
          description: Nurse not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
    put:
      tags:
        - Nurses
      summary: Update nurse
      operationId: updateNurse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNurseRequest"
      responses:
        "200":
          description: Nurse updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/Nurse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
    delete:
      tags:
        - Nurses
      summary: Delete nurse
      operationId: removeNurse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nurse deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
  /api/v1/patients:
    get:
      tags:
        - Patients
      summary: Get all patients
      operationId: getAllPatients
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      tags:
        - Patients
      summary: Create a new patient
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Patient created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Bad Request
  /api/v1/patients/{id}:
    get:
      tags:
        - Patients
      summary: Get patient by ID
      operationId: getPatientById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
    put:
      tags:
        - Patients
      summary: Update patient
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
      responses:
        "200":
          description: Patient updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "400":
          description: Bad Request
    delete:
      tags:
        - Patients
      summary: Delete patient
      operationId: removePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient deleted successfully
        "400":
          description: Bad Request
...
  /api/v1/vitalSign:
    post:
      tags:
        - VitalSign
      summary: Create a new vital sign
      operationId: createVitalSign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVitalSignRequest"
      responses:
        "201":
          description: Vital sign created successfully
  /api/v1/vitalSign/{id}:
    put:
      tags:
        - VitalSign
      summary: Update vital sign
      operationId: updateVitalSign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVitalSignRequest"
      responses:
        "200":
          description: Vital sign updated successfully
    delete:
      tags:
        - VitalSign
      summary: Delete vital sign
      operationId: removeVitalSign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vital sign deleted successfully
  /api/v1/vitalSign/appointment/{id}:
    get:
      tags:
        - VitalSign
      summary: Get vital sign by appointment ID
      operationId: getByAppointmentId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vital sign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalSign"
  /api/v1/medical-history:
    get:
      tags:
        - MedicalHistory
      summary: Get all medical histories
      operationId: getAllMedicalHistories
      responses:
        "200":
          description: List of medical histories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/MedicalHistory"
    post:
      tags:
        - MedicalHistory
      summary: Create a new medical history
      operationId: createMedicalHistory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalHistoryRequest"
      responses:
        "201":
          description: Medical history created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalHistory"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
  /api/v1/medical-history/patient/{patientId}:
    get:
      tags:
        - MedicalHistory
      summary: Get medical history by patient ID
      operationId: getMedicalHistoryByPatientId
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical history details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalHistory"
        "404":
          description: Medical history not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
    put:
      tags:
        - MedicalHistory
      summary: Update medical history
      operationId: updateMedicalHistory
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalHistoryRequest"
      responses:
        "200":
          description: Medical history updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/TResult"
                  - properties:
                      data:
                        $ref: "#/components/schemas/MedicalHistory"
    delete:
      tags:
        - MedicalHistory
      summary: Delete medical history
      operationId: removeMedicalHistory
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical history deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
...
components:
  schemas:
    TResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          nullable: true
...
    Nurse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        license:
          type: string
        officePhone:
          type: string
    CreateNurseRequest:
      type: object
      required:
        - email
        - password
        - name
        - LastName
        - phone
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        LastName:
          type: string
        phone:
          type: string
        license:
          type: string
        officePhone:
          type: string
    UpdateNurseRequest:
      type: object
      properties:
        name:
          type: string
        LastName:
          type: string
        phone:
          type: string
        license:
          type: string
        officePhone:
          type: string
    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          $ref: "#/components/schemas/Gender"
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePatientRequest:
      type: object
      required:
        - name
        - lastName
        - birthDate
        - gender
        - phone
        - email
        - address
      properties:
        name:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          $ref: "#/components/schemas/Gender"
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        isActive:
          type: boolean
    UpdatePatientRequest:
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          $ref: "#/components/schemas/Gender"
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        isActive:
          type: boolean
    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]
...
    VitalSign:
      type: object
      properties:
        id:
          type: integer
        temperature:
          type: number
        heartRate:
          type: integer
        bloodPressure:
          type: string
        oxygenSaturation:
          type: integer
        respiratoryRate:
          type: integer
        weight:
          type: number
        height:
          type: number
        appointmentId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateVitalSignRequest:
      type: object
      required:
        - temperature
        - heartRate
        - bloodPressure
        - oxygenSaturation
        - respiratoryRate
        - weight
        - height
        - appointmentId
      properties:
        temperature:
          type: number
        heartRate:
          type: integer
        bloodPressure:
          type: string
        oxygenSaturation:
          type: integer
        respiratoryRate:
          type: integer
        weight:
          type: number
        height:
          type: number
        appointmentId:
          type: integer
    UpdateVitalSignRequest:
      type: object
      properties:
        temperature:
          type: number
        heartRate:
          type: integer
        bloodPressure:
          type: string
        oxygenSaturation:
          type: integer
        respiratoryRate:
          type: integer
        weight:
          type: number
        height:
          type: number
    MedicalHistory:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        allergies:
          type: string
          nullable: true
        medications:
          type: string
          nullable: true
        surgeries:
          type: string
          nullable: true
        familyHistory:
          type: string
          nullable: true
        chronicConditions:
          type: string
          nullable: true
        gestationalHistory:
          type: string
          nullable: true
        reproductiveHistory:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateMedicalHistoryRequest:
      type: object
      required:
        - patientId
      properties:
        patientId:
          type: integer
        allergies:
          type: string
          nullable: true
        medications:
          type: string
          nullable: true
        surgeries:
          type: string
          nullable: true
        familyHistory:
          type: string
          nullable: true
        chronicConditions:
          type: string
          nullable: true
        gestationalHistory:
          type: string
          nullable: true
        reproductiveHistory:
          type: string
          nullable: true
    UpdateMedicalHistoryRequest:
      type: object
      properties:
        allergies:
          type: string
          nullable: true
        medications:
          type: string
          nullable: true
        surgeries:
          type: string
          nullable: true
        familyHistory:
          type: string
          nullable: true
        chronicConditions:
          type: string
          nullable: true
        gestationalHistory:
          type: string
          nullable: true
        reproductiveHistory:
          type: string
          nullable: true
...